+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug where `Components` attribute was not adding the correct traits
+
 ## [2.1.0.8] - 2025-03-18
 
 ### Added